Description: QT 模拟器中打开宏 ENABLE_MULTI_FONT IssueNo: https://gitee.com/openharmony/arkui_ui_lite/issues/I7TZPR Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: chenjiafeng <chenjiafeng10@huawei.com>

+++ frameworks/font/ui_font_vector.cpp
@@ -562,13 +562,13 @@ uint16_t UIFontVector::GetFontId(const char* ttfName, uint8_t fontSize) const
         int32_t i = 0;
         while (i < FONT_ID_MAX) {
             if ((fontInfo_[i].ttfName != nullptr) && (strstr(fontInfo_[i].ttfName, ttfName) != nullptr)) {
-                return static_cast<uint8_t>(i);
+                return static_cast<uint16_t>(i);
             }
             i++;
         }
     }
 
-    return FONT_ID_MAX;
+    return static_cast<uint16_t>(FONT_ID_MAX);
 }
 
 uint16_t UIFontVector::GetFontId(uint32_t unicode) const
